requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.0.6
django-filter==1.1.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-split-settings==0.3.0
djangorestframework==3.8.2
gunicorn==19.8.1
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.4
requests==2.19.1
simplejson==3.15.0
uritemplate==3.0.0
urllib3==1.23
whitenoise==3.3.1
